Refactored app to import Home page, and load user on mount.
